+++ blocks/init/src/blocks/components/hamburger/assets/index.js
@@ -1,0 +1,6 @@
+import { hamburgerInit, closeMobileMenuOnOverlayClick } from './hamburger';
+
+document.addEventListener('DOMContentLoaded', () => {
+  hamburgerInit();
+  closeMobileMenuOnOverlayClick();
+}, false);